requirements.txt
flask
flask-session
flask-cors
redis
bcrypt
mysql-connector-python
python-dotenv
openai
gunicorn
python-docx
